Fix missing epel-repo and file/folder permissions
The epel-repo is required for erlang to satisfy all dependencies.

Fix the permissions on /etc/rabbitmq/rabbitmq.config. This file needs to
be writable by any user as Openshift uses a random UID to run.
The guid of the user is always 0. So setting the files to group root and
setting write permissions to the files and folders fixes this issue.